Clean up texture readers (breaks cmap as texture support..)
