+++ include/ghoul/io/texture/texturereader.h
@@ -36,20 +36,37 @@
 
 namespace ghoul::io {
 
-class TextureReaderBase;
-
 /**
- * This class manages multiple TextureReaderBase and makes them available through one
- * method loadTexture. TextureReaderBases are added through the method addReader. The
- * class provides a static member, but also allows users to create local variants.
- * TextureReaderBases can be reused between multiple TextureReaders.
+ * A utility class for loading texture data from files or memory into OpenGL Texture
+ * objects using the STB image library. This class provides a centralized interface for
+ * loading textures from various common image formats and converting them into GPU-ready
+ * texture objects.
  *
- * \see TextureReaderBase
+ * The class uses the singleton pattern through its static `ref()` method, providing a
+ * globally accessible instance.
+ *
+ * The class supports loading textures with 1, 2, or 3 dimensions and provides
+ * functionality to:
+ * - Load textures directly from file paths
+ * - Load textures from raw memory buffers
+ * - Query image dimensions without full texture creation
+ * - Configure texture sampler settings (filtering, wrapping, etc.)
+ *
+ * Supported file formats include:
+ * - JPEG (.jpeg, .jpg)
+ * - PNG (.png)
+ * - BMP (.bmp)
+ * - TGA (.tga)
+ * - PSD (.psd)
+ * - GIF (.gif)
+ * - HDR (.hdr)
+ * - PIC (.pic)
+ * - PNM (.ppm, .pgm)
  */
 class TextureReader {
 public:
     /**
-     * Exception that gets thrown when there is no reader for the provided \p extension.
+     * Exception that gets thrown when the provided \p extension is not supported.
      */
     struct MissingReaderException final : public RuntimeError {
         explicit MissingReaderException(std::string extension,
@@ -57,6 +74,19 @@ public:
 
         const std::string fileExtension;
         const std::filesystem::path file;
+    };
+
+    /**
+     * The exception that gets thrown if there was an error loading the Texture.
+     */
+    struct TextureLoadException final : public RuntimeError {
+        explicit TextureLoadException(std::filesystem::path name, std::string msg);
+
+        /// The filename that caused the exception to be thrown
+        const std::filesystem::path filename;
+
+        /// The error message that occurred
+        const std::string errorMessage;
     };
 
     /**
@@ -77,8 +107,8 @@ public:
     static TextureReader& ref();
 
     /**
-     * Loads the provided \p filename into a Texture and returns it. The correct
-     * TextureReaderBase is determined by the extension of the \p filename.
+     * Loads the provided \p filename into a Texture and returns it. The image format is
+     * determined by the extension of the \p filename.
      *
      * \param filename The name of the file which should be loaded into a texture
      * \param nDimensions The number of dimensions of the texture that are returned when
@@ -90,23 +120,17 @@ public:
      *        created from the contents of the \p filename
      *
      * \throw TextureLoadException If there was an error reading the \p filename
-     * \throw MissingReaderException If there was no reader for the specified \p filename
+     * \throw MissingReaderException If the extension in the \p filename is not supported
      * \pre \p filename must not be empty
      * \pre \p filename must have an extension
      * \pre \p nDimensions The number of texture dimension must be 1, 2, or 3
-     * \pre At least one TextureReaderBase must have been added to the TextureReader
-     *      before (addReader)
      */
     std::unique_ptr<opengl::Texture> loadTexture(const std::filesystem::path& filename,
         int nDimensions, opengl::Texture::SamplerInit samplerSettings = {});
 
     /**
      * Loads a Texture from the memory pointed at by \p memory. The memory block must
-     * contain at least \p size number of bytes. The bytes are then passed to the actual
-     * implementation of the texture reader to create the Textjure object. The optional
-     * \p format parameter is used to disambiguate the cases where multiple TextureReaders
-     * are registered. In this case, the \p format is used in the same way as the file
-     * extension for the #loadTexture method.
+     * contain at least \p size number of bytes.
      *
      * \param memory The memory that contains the bytes of the Texture to be loaded
      * \param size The number of bytes contained in \p memory
@@ -116,14 +140,13 @@ public:
      *        example, someone might want to load a 128x1 texture but use it as a 2D
      *        texture instead
      * \param format The format of the image pointed to by \p memory. This parameter
-     *        should be the same as the usual file extension for the image. However, this
-     *        parameter is only used to determine which TextureReader is used for this
-     *        memory, if multiple readers are registered
+     *        should be the same as the usual file extension for the image and is
+     *        used to determine if the file type is supported by the reader
      * \param samplerSettings The settings that should be used for the Texture that is
      *        created from the contents of the \p memory
      *
      * \throw TextureLoadException If there was an error reading the \p memory
-     * \throw MissingReaderException If there was no reader for the specified \p filename
+     * \throw MissingReaderException If the extension in the \p filename is not supported
      * \pre \p memory must not be `nullptr`
      * \pre \p size must be > 0
      * \pre \p nDimensions The number of texture dimension must be 1, 2, or 3
@@ -141,42 +164,25 @@ public:
      * \throw TextureLoadException If there was an error loading the texture
      * \pre \p filename must not be empty
      * \pre The extension of \p filename must be among the supported extensions as
-     *      reported by supportedExtensions
+     *      reported by `supportedExtensions`
      */
     glm::ivec2 imageSize(const std::filesystem::path& filename) const;
 
     /**
-     * Returns a list of all the extensions that are supported by registered readers. If a
-     * file with an extension included in this list is passed to the loadTexture file and
-     * the file is not corrupted, it will be successfully loaded.
+     * Returns Whether the provided file \p extension is supported by this reader.
+     *
+     * \return True if the provided \p extension is supported, false otherwise
+     */
+    bool isSupportedExtension(const std::string& extension) const;
+
+    /**
+     * Returns a list of all the extensions that are supported. If a file with an
+     * extension included in this list is passed to the loadTexture file and the file
+     * is not corrupted, it will be successfully loaded.
      *
      * \return A list of all supported extensions
      */
-    std::vector<std::string> supportedExtensions();
-
-    /**
-     * Adds the \p reader to this TextureReader and makes it available through subsequent
-     * calls to loadTexture. If an extension is supported by multiple TextureReaderBases,
-     * the TextureReaderBase that was added first will be used.
-     *
-     * \param reader The reader that is to be added to this TextureReader
-     *
-     * \pre \p reader must not have been added to this TextureReader before
-     */
-    void addReader(std::unique_ptr<TextureReaderBase> reader);
-
-private:
-    /**
-     * Returns the TextureReaderBase that is responsible for the provided extension.
-     *
-     * \param extension The extension for which the TextureReaderBase should be returned
-     * \return The first match of a reader that can read the provided \p extension, or
-     *         `nullptr` if no such reader exists
-     */
-    TextureReaderBase* readerForExtension(const std::string& extension) const;
-
-    /// The list of all registered readers
-    std::vector<std::unique_ptr<TextureReaderBase>> _readers;
+    std::vector<std::string> supportedExtensions() const;
 };
 
 } // namespace ghoul::io
