+++ src/CMakeLists.txt
@@ -235,9 +235,6 @@ if (GHOUL_MODULE_OPENGL)
       ${PROJECT_SOURCE_DIR}/include/ghoul/opengl/textureunit.h
       ${PROJECT_SOURCE_DIR}/include/ghoul/opengl/uniformcache.h
       ${PROJECT_SOURCE_DIR}/include/ghoul/io/texture/texturereader.h
-      ${PROJECT_SOURCE_DIR}/include/ghoul/io/texture/texturereaderbase.h
-      ${PROJECT_SOURCE_DIR}/include/ghoul/io/texture/texturereadercmap.h
-      ${PROJECT_SOURCE_DIR}/include/ghoul/io/texture/texturereaderstb.h
       ${PROJECT_SOURCE_DIR}/include/ghoul/io/texture/texturewriter.h
       opengl/debugcontext.cpp
       opengl/framebufferobject.cpp
@@ -252,9 +249,6 @@ if (GHOUL_MODULE_OPENGL)
       opengl/texturemanager.cpp
       opengl/textureunit.cpp
       io/texture/texturereader.cpp
-      io/texture/texturereaderbase.cpp
-      io/texture/texturereadercmap.cpp
-      io/texture/texturereaderstb.cpp
       io/texture/texturewriter.cpp
   )
 endif ()
