--- src/io/texture/texturereaderbase.cpp
@@ -1,42 +1,0 @@
-/*****************************************************************************************
- *                                                                                       *
- * GHOUL                                                                                 *
- * General Helpful Open Utility Library                                                  *
- *                                                                                       *
- * Copyright (c) 2012-2026                                                               *
- *                                                                                       *
- * Permission is hereby granted, free of charge, to any person obtaining a copy of this  *
- * software and associated documentation files (the "Software"), to deal in the Software *
- * without restriction, including without limitation the rights to use, copy, modify,    *
- * merge, publish, distribute, sublicense, and/or sell copies of the Software, and to    *
- * permit persons to whom the Software is furnished to do so, subject to the following   *
- * conditions:                                                                           *
- *                                                                                       *
- * The above copyright notice and this permission notice shall be included in all copies *
- * or substantial portions of the Software.                                              *
- *                                                                                       *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED,   *
- * INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A         *
- * PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT    *
- * HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF  *
- * CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE  *
- * OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.                                         *
- ****************************************************************************************/
-
-#include <ghoul/io/texture/texturereaderbase.h>
-
-#include <ghoul/format.h>
-#include <utility>
-
-namespace ghoul::io {
-
-TextureReaderBase::TextureLoadException::TextureLoadException(std::filesystem::path name,
-                                                              std::string msg,
-                                                              const TextureReaderBase* r)
-    : RuntimeError(std::format("Error loading texture '{}'", name), "TextureLoader")
-    , filename(std::move(name))
-    , errorMessage(std::move(msg))
-    , reader(r)
-{}
-
-} // namespace ghoul::io
